chore: reworked logger init
The whole logger setup is kinda silly, and was primarially silly
because a logger was not initialized until after the config was
parsed for debug handling.

Now setting the logger to a default value, and redefine it later once config
is parsed.

There is also no reason to call initLogger and initConfig in additional
commands with this change.